apps/fireedge/pyproject.toml
[project]
name = "fireedge"
version = "0.1.0"
description = "Smoke-penetrating wildfire detection on satellite edge using LFM 2.5-VL"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.26",
    "pillow>=10.0",
    "requests>=2.31",
    "torch>=2.3",
    "torchvision>=0.18",
    "transformers>=4.45",
    "accelerate>=0.33",
    "earthaccess>=0.14.0",
    "netcdf4>=1.7.3",
    "peft>=0.19.1",
    "trl>=1.1.0",
    "datasets>=4.8.4",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "jupyter>=1.0",
    "matplotlib>=3.8",
    "ipywidgets>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "notebook>=7.5.5",
    "pytest>=9.0.3",
]
